% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation2.R
\docType{package}
\name{aggregation2}
\alias{aggregation2}
\alias{aggregation2-package}
\title{Aggregation 2 experiment}
\description{
Data, materials, and experiment code for the Aggregation 2 experiment.
}
\details{
aggregation2
A data package for the Aggregation 2 experiment.

Use \code{data(package='aggregation2')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{data}
\link{power}
}
