% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_joint.R
\name{clean_data_joint}
\alias{clean_data_joint}
\title{Clean joint data}
\usage{
clean_data_joint(data_raw_prep, names_to)
}
\arguments{
\item{data_raw_prep}{}

\item{names_to}{}
}
\value{

}
\description{
Clean joint data
}
\author{
Shir Dekel
}
