% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_separate.R
\name{clean_data_separate}
\alias{clean_data_separate}
\title{Clean separate data}
\usage{
clean_data_separate(data_raw_prep, names_to)
}
\arguments{
\item{data_raw_prep}{}

\item{names_to}{}
}
\value{

}
\description{
Clean separate data
}
\author{
Shir Dekel
}
