% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_name_materials_experiment2.R
\name{get_file_name_materials_experiment2}
\alias{get_file_name_materials_experiment2}
\title{Get E2 screenshot file names}
\usage{
get_file_name_materials_experiment2()
}
\value{

}
\description{
Get E2 screenshot file names
}
\author{
Shir Dekel
}
