% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_preamble_distribution.R
\name{get_preamble_distribution}
\alias{get_preamble_distribution}
\title{Get distribution preamble}
\usage{
get_preamble_distribution()
}
\value{

}
\description{
Get distribution preamble
}
\author{
Shir Dekel
}
