% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_description_experiment2.R
\name{get_project_description_experiment2}
\alias{get_project_description_experiment2}
\title{Get project description E2}
\usage{
get_project_description_experiment2(gambles, project_name, project_type)
}
\arguments{
\item{project_type}{}
}
\value{

}
\description{
Get project description E2
}
\author{
Shir Dekel
}
