% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projects_experiment2.R
\name{get_projects_experiment2}
\alias{get_projects_experiment2}
\title{Get projects with short descriptions}
\usage{
get_projects_experiment2(gambles)
}
\arguments{
\item{gambles}{}
}
\value{

}
\description{
Get projects with short descriptions
}
\author{
Shir Dekel
}
