% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_questions_joint.R
\name{get_questions_joint}
\alias{get_questions_joint}
\title{Joint questions}
\usage{
get_questions_joint(
  project_description,
  form_options = c("Yes", "No"),
  project_input
)
}
\arguments{
\item{project_description}{}

\item{project_input}{}
}
\value{

}
\description{
Joint questions
}
\author{
Shir Dekel
}
