% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_restriction_values.R
\name{get_restriction_values}
\alias{get_restriction_values}
\title{Get restriction values (EV and gain/loss)}
\usage{
get_restriction_values(prob_positive_sample, outcome_positive_sample)
}
\value{

}
\description{
Get restriction values (EV and gain/loss)
}
\author{
Shir Dekel
}
