% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trial_awareness.R
\name{get_trial_awareness}
\alias{get_trial_awareness}
\title{Get awareness trial}
\usage{
get_trial_awareness(stimulus, awareness)
}
\arguments{
\item{stimulus}{}

\item{awareness}{}
}
\value{

}
\description{
Get awareness trial
}
\author{
Shir Dekel
}
