% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trial_joint.R
\name{get_trial_joint}
\alias{get_trial_joint}
\title{Get joint trial}
\usage{
get_trial_joint(
  preamble = "<p>Indicate below whether you would invest in the following:</p>",
  questions_joint,
  distribution,
  randomize_order
)
}
\arguments{
\item{randomize_order}{}
}
\value{

}
\description{
Get joint trial
}
\author{
Shir Dekel
}
