% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation2.R
\docType{data}
\name{power}
\alias{power}
\title{Descriptive data title}
\format{
a \code{numeric} containing the following fields:
\describe{
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{aggregation2}
}
